SAK-10090 http://jira.sakaiproject.org/jira/browse/SAK-10090 gradebook setup/ running total is displaying truncated version of %'s entered
git-svn-id: https://source.sakaiproject.org/svn/gradebook/trunk@30875 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ gradebook/app/ui/src/webapp/js/gradebook.js
@@ -163,7 +163,7 @@ function updateRunningTotal(thisForm) {
     var highlightTotal = true;
 
     if (elementNamePieces[3] == "weightInput") {
-        weight = parseInt(formElement.value);
+        weight = parseFloat(formElement.value);
 
         if (weight >= 0) {
             runningTotal += weight;
